[PSEC-1106] Split cloudtrail client for new auditing task
