+++ tests/clients/test_aws_client_factory.py
@@ -235,7 +235,6 @@ class TestGetClients(TestCase):
     def test_get_cloudtrail_client(self, _: Mock) -> None:
         cloudtrail_boto_client = Mock()
         logs_client = Mock()
-        s3_client = Mock()
         with patch(
             f"{self.factory_path}.get_cloudtrail_boto_client",
             side_effect=lambda acc: cloudtrail_boto_client if acc == account() else None,
@@ -244,14 +243,9 @@ class TestGetClients(TestCase):
                 f"{self.factory_path}.get_logs_client",
                 side_effect=lambda acc: logs_client if acc == account() else None,
             ):
-                with patch(
-                    f"{self.factory_path}.get_s3_client",
-                    side_effect=lambda acc: s3_client if acc == account() else None,
-                ):
-                    cloudtrail_client = AwsClientFactory(self.mfa, self.username).get_cloudtrail_client(account())
-                    self.assertEqual(cloudtrail_client._cloudtrail, cloudtrail_boto_client)
-                    self.assertEqual(cloudtrail_client._logs, logs_client)
-                    self.assertEqual(cloudtrail_client._s3, s3_client)
+                cloudtrail_client = AwsClientFactory(self.mfa, self.username).get_cloudtrail_client(account())
+                self.assertEqual(cloudtrail_client._cloudtrail, cloudtrail_boto_client)
+                self.assertEqual(cloudtrail_client._logs, logs_client)
 
 
 @patch.object(AwsClientFactory, "_get_session_token")
