+++ tests/clients/test_aws_cloudtrail_client.py
@@ -2,13 +2,13 @@ from unittest.mock import Mock
 
 from src.clients.composite.aws_cloudtrail_client import AwsCloudtrailClient
 
-from tests.test_types_generator import log_group, bucket
+from tests.test_types_generator import log_group
 
 import tests.clients.test_aws_cloudtrail_responses as resp
 
 
-def client(cloudtrail: Mock = Mock(), logs: Mock = Mock(), s3: Mock = Mock()) -> AwsCloudtrailClient:
-    return AwsCloudtrailClient(cloudtrail, logs, s3)
+def client(cloudtrail: Mock = Mock(), logs: Mock = Mock()) -> AwsCloudtrailClient:
+    return AwsCloudtrailClient(cloudtrail, logs)
 
 
 def test_get_trails_success() -> None:
@@ -43,17 +43,3 @@ def test_get_cloudtrail_log_group() -> None:
 def test_get_cloudtrail_log_group_not_found() -> None:
     logs = Mock(describe_log_groups=Mock(return_value=[]))
     assert client(logs=logs).get_cloudtrail_log_group() is None
-
-
-def test_get_event_bucket() -> None:
-    the_policy = {"banana": 1}
-    s3_client = Mock(
-        get_bucket_policy=Mock(side_effect=lambda b: the_policy if b == "cloudtrail-logs-bucket" else None)
-    )
-    actual_bucket = client(s3=s3_client).get_event_bucket()
-    assert actual_bucket == bucket(name="cloudtrail-logs-bucket", policy=the_policy)
-
-
-def test_get_event_bucket_does_not_exist() -> None:
-    s3_client = Mock(get_bucket_policy=Mock(return_value=None))
-    assert client(s3=s3_client).get_event_bucket() is None
